+++ app/src/main/java/com/mendosal/locationsweather/data/remote/models/Main.kt
@@ -1,10 +1,10 @@
 package com.mendosal.locationsweather.data.remote.models
 
 data class Main(
-    val feels_like: Double,
+    val feels_like: Float,
     val humidity: Int,
     val pressure: Int,
-    val temp: Double,
-    val temp_max: Double,
-    val temp_min: Double
+    val temp: Float,
+    val temp_max: Float,
+    val temp_min: Float
 )
