+++ app/src/main/java/com/mendosal/locationsweather/data/remote/models/Wind.kt
@@ -2,5 +2,5 @@ package com.mendosal.locationsweather.data.remote.models
 
 data class Wind(
     val deg: Int,
-    val speed: Double
+    val speed: Float
 )
